Integ 3.6/message
 - HTTP 0.9 support
 - manage www-urlencoded content, and plain/text content
 - fix few parsing errors

+++ include/ouistiti/httpserver.h
@@ -90,7 +90,7 @@ typedef enum
 	HTTP_PIPELINE = 0x0100,
 } http_message_version_e;
 
-EXPORT_SYMBOL int httpserver_version(http_message_version_e versionid, const char **version);
+EXPORT_SYMBOL size_t httpserver_version(http_message_version_e versionid, const char **version);
 
 typedef int http_message_result_e;
 #define RESULT_200 200
@@ -582,7 +582,7 @@ EXPORT_SYMBOL const char * httpmessage_SERVER(http_message_t *message, const cha
  * @return the value of the attribut or a empty string
  */
 EXPORT_SYMBOL const char * httpmessage_REQUEST(http_message_t *message, const char *key);
-EXPORT_SYMBOL int httpmessage_REQUEST2(http_message_t *message, const char *key, const char **value);
+EXPORT_SYMBOL size_t httpmessage_REQUEST2(http_message_t *message, const char *key, const char **value);
 
 /**
  * @brief get/set value for the session used by the request
